Convert a CSV into a serial mapping suitable for Contextractr

Given a csv, with one row per keyword + approx.match specifier, grouped according to "title" (Single value for title may span multiple keywords) transform this to something digestable by contextractr.

csv_to_serial(df)

Arguments

df
the input dataframe, more or less a raw csv

Value

the input df, serialized.

Examples

if (requireNamespace("readr", quietly = TRUE)){
  in_csv <- readr::read_csv("../../tests/testthat/csv_specifier.csv")
  print(in_csv)

  out <- in_csv %>% csv_to_serial()
  if (requireNamespace("utils", quietly = TRUE)) utils::str(out)
}

#> Parsed with column specification:
#> cols(
#>   title = col_character(),
#>   keywords = col_character(),
#>   approx.match = col_logical(),
#>   comments = col_character()
#> )

#> # A tibble: 40 x 4
#>    title      keywords   approx.match comments                                  
#>    <chr>      <chr>      <lgl>        <chr>                                     
#>  1 A2RB       telmisart… NA           NA                                        
#>  2 A2RB       micardis   NA           NA                                        
#>  3 ACE Inhib… perindopr… NA           NA                                        
#>  4 ACE Inhib… coversyl   NA           NA                                        
#>  5 aspirin    aspirin    NA           Match aspirin with aspirin. If approx.mat…
#>  6 aspirin    spren      NA           Comments are ignored, just add your thoug…
#>  7 aspirin    cartia     NA           Aspirin should match aspirin, spren, or c…
#>  8 Beta bloc… SOTALOL    NA           NA                                        
#>  9 Beta bloc… metoprolol NA           NA                                        
#> 10 Frusemide… frusemide  NA           NA                                        
#> # … with 30 more rows
#> List of 17
#>  $ :List of 3
#>   ..$ title       : chr "A2RB"
#>   ..$ keywords    : chr [1:2] "telmisartan" "micardis"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "ACE Inhibitor"
#>   ..$ keywords    : chr [1:2] "perindopril" "coversyl"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "aspirin"
#>   ..$ keywords    : chr [1:3] "aspirin" "spren" "cartia"
#>   ..$ approx.match: num [1:3] 0 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Beta blocker"
#>   ..$ keywords    : chr [1:2] "SOTALOL" "metoprolol"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Frusemide/Lasix"
#>   ..$ keywords    : chr [1:2] "frusemide" "lasix"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "insulin"
#>   ..$ keywords    : chr "PROTOPHANE"
#>   ..$ approx.match: num 0
#>  $ :List of 3
#>   ..$ title       : chr "Metformin"
#>   ..$ keywords    : chr "metformin"
#>   ..$ approx.match: num 0
#>  $ :List of 3
#>   ..$ title       : chr "Nitrate"
#>   ..$ keywords    : chr [1:6] "ISOSORBIDE MONONITRATE" "NICORANDIL" "GLYCERYL TRINITRATE" "nicorandil" ...
#>   ..$ approx.match: num [1:6] 0 0 0 0 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Other anticoagulant"
#>   ..$ keywords    : chr [1:5] "heparin" "apixaban" "eliquis" "rivaroxaban" ...
#>   ..$ approx.match: num [1:5] 0 0 0 0 0
#>  $ :List of 3
#>   ..$ title       : chr "other antihypertensive"
#>   ..$ keywords    : chr "HYDROCHLOROTHIAZIDE"
#>   ..$ approx.match: num 0
#>  $ :List of 3
#>   ..$ title       : chr "Other antiplatelet"
#>   ..$ keywords    : chr [1:2] "clopidogrel" "plavix"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Other diabetic drug"
#>   ..$ keywords    : chr [1:2] "DAPAGLIFAZOLIN" "EMPAGLIFLOZIN"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Other lipid lowering"
#>   ..$ keywords    : chr "FENOFIBRATE"
#>   ..$ approx.match: num 0
#>  $ :List of 3
#>   ..$ title       : chr "statin"
#>   ..$ keywords    : chr [1:2] "EZETIMIBE" "ROSUVASTATIN"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Statin"
#>   ..$ keywords    : chr [1:2] "atorvastatin" "lipitor"
#>   ..$ approx.match: num [1:2] 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Vitamin supplementation"
#>   ..$ keywords    : chr [1:4] "multivitamin" "folic acid" "fish oil" "inner health plus"
#>   ..$ approx.match: num [1:4] 0 0 0 0
#>  $ :List of 3
#>   ..$ title       : chr "Warfarin"
#>   ..$ keywords    : chr [1:2] "warfarin" "coumadin"
#>   ..$ approx.match: num [1:2] 0 0